JS-135: Upgrade commons-httpclient from 1.3 to httpcore 1.4 and httpclient 1.4-beta1 (requires Java 5) Note this significantly hit WebResource and now shutdown needs to be called on it.

+++ jsword-limbo/src/main/java/org/crosswire/jsword/book/install/sword/FtpSwordInstaller.java
@@ -61,7 +61,7 @@ public class FtpSwordInstaller extends AbstractSwordInstaller {
      * org.crosswire.jsword.book.install.Installer#getSize(org.crosswire.jsword
      * .book.Book)
      */
-    public int getSize(Book book) {
+    public long getSize(Book book) {
         // not implemented
         return 0;
     }
